Fixes blank "show tag values" on influxdb v1.6
I'm stuck on influxdb v1.6 as running this on a pi using the default influxdb in the repo. The "show tag values" query does not return anything on older versions of influx db. New query works on older versions where the bug is present and shouldn't hit performance too badly.